Maybe CMake

+++ include/globals.hpp
@@ -2,6 +2,6 @@
 
 #include <memory>
 
-extern uint64_t CELL_AMOUNT;
+extern unsigned int CELL_AMOUNT;
 extern std::shared_ptr<unsigned char[]> cells;
-extern uint64_t pointer;
+extern unsigned int pointer;
